Adapt Email Templaes to only allow creation and update, and add first simmple UI to fetch sort and chategorize the Templates

+++ server/src/main/java/de/tum/cit/aet/thesis/service/EmailTemplateService.java
@@ -2,6 +2,7 @@ package de.tum.cit.aet.thesis.service;
 
 import de.tum.cit.aet.thesis.entity.EmailTemplate;
 import de.tum.cit.aet.thesis.entity.ResearchGroup;
+import de.tum.cit.aet.thesis.exception.request.AccessDeniedException;
 import de.tum.cit.aet.thesis.exception.request.ResourceNotFoundException;
 import de.tum.cit.aet.thesis.repository.EmailTemplateRepository;
 import de.tum.cit.aet.thesis.repository.ResearchGroupRepository;
@@ -17,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
+import java.util.Set;
 import java.util.UUID;
 
 @Service
@@ -37,6 +39,35 @@ public class EmailTemplateService {
         return currentUserProviderProvider.getObject();
     }
 
+    private static final Set<String> VALID_TEMPLATE_CASES = Set.of(
+            "APPLICATION_REJECTED_TOPIC_REQUIREMENTS",
+            "THESIS_PRESENTATION_UPDATED",
+            "THESIS_ASSESSMENT_ADDED",
+            "THESIS_PRESENTATION_SCHEDULED",
+            "THESIS_CREATED",
+            "APPLICATION_REJECTED_TOPIC_OUTDATED",
+            "THESIS_FINAL_SUBMISSION",
+            "THESIS_PROPOSAL_REJECTED",
+            "THESIS_COMMENT_POSTED",
+            "THESIS_PRESENTATION_DELETED",
+            "THESIS_PROPOSAL_ACCEPTED",
+            "APPLICATION_REJECTED",
+            "APPLICATION_REJECTED_TITLE_NOT_INTERESTING",
+            "APPLICATION_ACCEPTED",
+            "APPLICATION_CREATED_CHAIR",
+            "APPLICATION_ACCEPTED_NO_ADVISOR",
+            "APPLICATION_REJECTED_TOPIC_FILLED",
+            "THESIS_CLOSED",
+            "THESIS_PRESENTATION_INVITATION_UPDATED",
+            "APPLICATION_CREATED_STUDENT",
+            "THESIS_PRESENTATION_INVITATION_CANCELLED",
+            "THESIS_PRESENTATION_INVITATION",
+            "THESIS_PROPOSAL_UPLOADED",
+            "APPLICATION_REMINDER",
+            "THESIS_FINAL_GRADE",
+            "APPLICATION_REJECTED_STUDENT_REQUIREMENTS"
+    );
+
     public Page<EmailTemplate> getAll(
             String[] templateCases,
             String[] languages,
@@ -44,11 +75,15 @@ public class EmailTemplateService {
             int page,
             int limit,
             String sortBy,
-            String sortOrder
+            String sortOrder,
+            UUID researchGroupId
     ) {
+        if (!currentUserProvider().isAdmin() && !currentUserProvider().getResearchGroupOrThrow().getId().equals(researchGroupId) && researchGroupId != null ) {
+            throw new AccessDeniedException("You do not have access to this research group.");
+        }
 
-        UUID researchGroupId = currentUserProvider().isAdmin() ?
-                null : currentUserProvider().getResearchGroupOrThrow().getId();
+        UUID searchResearchGroupId = currentUserProvider().isAdmin() ? researchGroupId : currentUserProvider().getResearchGroupOrThrow().getId();
+
         String searchQueryFilter =
                 searchQuery == null || searchQuery.isEmpty() ? null : searchQuery.toLowerCase();
         String[] templateCasesFilter = templateCases == null || templateCases.length == 0 ? null : templateCases;
@@ -64,7 +99,7 @@ public class EmailTemplateService {
                 : PageRequest.of(page, limit, Sort.by(order));
 
         return emailTemplateRepository.searchEmailTemplate(
-                researchGroupId,
+                searchResearchGroupId,
                 templateCasesFilter,
                 languagesFilter,
                 searchQueryFilter,
@@ -94,17 +129,27 @@ public class EmailTemplateService {
             researchGroup = researchGroupRepository.findById(researchGroupId)
                     .orElseThrow(() -> new ResourceNotFoundException(
                             String.format("Research Group with id %s not found.", researchGroupId)));
+        } else if (!currentUserProvider().isAdmin()) { // Allow only admins to create default templates
+            throw new AccessDeniedException("Only admins can create default email templates without ResearchGroupId.");
         }
-        currentUserProvider().assertCanAccessResearchGroup(researchGroup);
 
-        EmailTemplate emailTemplate = new EmailTemplate();
+        currentUserProvider().assertCanAccessResearchGroup(researchGroup);
+        validateTemplateCase(templateCase);
+
+        // Check if a template with the same case and research group already exists -> update instead of create
+        EmailTemplate emailTemplate = emailTemplateRepository
+                .findByResearchGroupAndTemplateCase(researchGroupId, templateCase)
+                .orElse(new EmailTemplate());
+
         emailTemplate.setTemplateCase(templateCase);
         emailTemplate.setDescription(description);
         emailTemplate.setSubject(subject);
         emailTemplate.setBodyHtml(bodyHtml);
         emailTemplate.setLanguage(language);
         emailTemplate.setResearchGroup(researchGroup);
-        emailTemplate.setCreatedAt(Instant.now());
+        if (emailTemplate.getId() == null) {
+            emailTemplate.setCreatedAt(Instant.now()); //only set createdAt if it's a new template
+        }
         emailTemplate.setUpdatedAt(Instant.now());
         emailTemplate.setUpdatedBy(currentUserProvider().getUser());
 
@@ -121,6 +166,12 @@ public class EmailTemplateService {
             String language
     ) {
         currentUserProvider().assertCanAccessResearchGroup(emailTemplate.getResearchGroup());
+        validateTemplateCase(templateCase);
+
+        if (emailTemplate.getResearchGroup() == null && !currentUserProvider().isAdmin()) {
+            throw new AccessDeniedException("Only admins can update default email templates without ResearchGroupId.");
+        }
+
         emailTemplate.setTemplateCase(templateCase);
         emailTemplate.setDescription(description);
         emailTemplate.setSubject(subject);
@@ -137,8 +188,19 @@ public class EmailTemplateService {
         EmailTemplate emailTemplate = emailTemplateRepository.findById(emailTemplateId)
                 .orElseThrow(() -> new ResourceNotFoundException(
                         String.format("Email Template with id %s not found.", emailTemplateId)));
+
+        if (emailTemplate.getResearchGroup() == null && !currentUserProvider().isAdmin()) {
+            throw new AccessDeniedException("Only admins can delete default email templates without ResearchGroupId.");
+        }
+
         currentUserProvider().assertCanAccessResearchGroup(emailTemplate.getResearchGroup());
 
         emailTemplateRepository.deleteById(emailTemplateId);
+    }
+
+    private void validateTemplateCase(String templateCase) {
+        if (!VALID_TEMPLATE_CASES.contains(templateCase)) {
+            throw new IllegalArgumentException("Invalid template case: " + templateCase);
+        }
     }
 }
